Simplify feature interface
