+++ python/ray/dashboard/client/src/pages/dashboard/features/Disk.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { formatUsage } from "../../../common/formatUtils";
 import UsageBar from "../../../common/UsageBar";
-import { makeFeature } from "./Feature";
+import { makeFeature } from "./makeFeature";
 import Typography from "@material-ui/core/Typography";
 
 const Disk = makeFeature({
