+++ python/ray/dashboard/client/src/pages/dashboard/features/Host.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { makeFeature } from "./Feature";
+import { makeFeature } from "./makeFeature";
 
 const Host = makeFeature({
   getFeatureForNode: ({ node }) => (
