+++ python/ray/dashboard/client/src/pages/dashboard/features/Logs.tsx
@@ -2,7 +2,7 @@ import Link from "@material-ui/core/Link";
 import Typography from "@material-ui/core/Typography";
 import React from "react";
 import { Link as RouterLink } from "react-router-dom";
-import { makeFeature } from "./Feature";
+import { makeFeature } from "./makeFeature";
 
 export const makeLogsFeature = (logCounts: {
   perWorker: { [pid: string]: number };
